Here is a code for parsing from web-site santech.ru
I've used the BeatifulSoup for parsing html-code and finding through the web-site link that I need, and download them, and then via Pandas.DataFrame make an excel-table with all necesery data that I need. Also, I make a document parser, which downloading documents, but sometimes the documents are the similar, so this code making a folders and downloading only new documents and delete empty folders.